Remove config interface. Make CreateImage function to accept config as list of config fields
