+++ builder/googlecompute/config_test.go
@@ -362,16 +362,16 @@ func TestConfigPrepareStartupScriptFile(t *testing.T) {
 
 func TestConfigDefaults(t *testing.T) {
 	cases := []struct {
-		Read  func(c *ConfigGCE) interface{}
+		Read  func(c *Config) interface{}
 		Value interface{}
 	}{
 		{
-			func(c *ConfigGCE) interface{} { return c.Comm.Type },
+			func(c *Config) interface{} { return c.Comm.Type },
 			"ssh",
 		},
 
 		{
-			func(c *ConfigGCE) interface{} { return c.Comm.SSHPort },
+			func(c *Config) interface{} { return c.Comm.SSHPort },
 			22,
 		},
 	}
@@ -437,7 +437,7 @@ func testConfig(t *testing.T) (config map[string]interface{}, tempAccountFile st
 	return config, tempAccountFile
 }
 
-func testConfigStruct(t *testing.T) *ConfigGCE {
+func testConfigStruct(t *testing.T) *Config {
 	raw, tempfile := testConfig(t)
 	defer os.Remove(tempfile)
