+++ builder/googlecompute/step_check_existing_image_test.go
@@ -18,7 +18,7 @@ func TestStepCheckExistingImage(t *testing.T) {
 
 	state.Put("instance_name", "foo")
 
-	config := state.Get("config").(*ConfigGCE)
+	config := state.Get("config").(*Config)
 	driver := state.Get("driver").(*DriverMock)
 	driver.ImageExistsResult = true
