+++ builder/googlecompute/step_create_image.go
@@ -19,7 +19,7 @@ type StepCreateImage int
 // The image is created from the persistent disk used by the instance. The
 // instance must be deleted and the disk retained before doing this step.
 func (s *StepCreateImage) Run(ctx context.Context, state multistep.StateBag) multistep.StepAction {
-	config := state.Get("config").(*ConfigGCE)
+	config := state.Get("config").(*Config)
 	driver := state.Get("driver").(Driver)
 	ui := state.Get("ui").(packer.Ui)
 
@@ -38,7 +38,9 @@ func (s *StepCreateImage) Run(ctx context.Context, state multistep.StateBag) mul
 
 	ui.Say("Creating image...")
 
-	imageCh, errCh := driver.CreateImage(config)
+	imageCh, errCh := driver.CreateImage(
+		config.ImageName, config.ImageDescription, config.ImageFamily, config.Zone,
+		config.DiskName, config.ImageLabels, config.ImageLicenses, config.ImageEncryptionKey)
 	var err error
 	select {
 	case err = <-errCh:
