+++ builder/googlecompute/step_wait_startup_script_test.go
@@ -11,7 +11,7 @@ import (
 func TestStepWaitStartupScript(t *testing.T) {
 	state := testState(t)
 	step := new(StepWaitStartupScript)
-	c := state.Get("config").(*ConfigGCE)
+	c := state.Get("config").(*Config)
 	d := state.Get("driver").(*DriverMock)
 
 	testZone := "test-zone"
