+++ builder/googlecompute/winrm.go
@@ -7,7 +7,7 @@ import (
 
 // winrmConfig returns the WinRM configuration.
 func winrmConfig(state multistep.StateBag) (*communicator.WinRMConfig, error) {
-	config := state.Get("config").(*ConfigGCE)
+	config := state.Get("config").(*Config)
 	password := state.Get("winrm_password").(string)
 
 	return &communicator.WinRMConfig{
